feat(dashboard/recording): add horizontal quick action carousel & hospital pen management
- Implement unified horizontal scroll carousel on dashboard containing Pencatatan Keuangan, Sekat Seleksian, Monitor FCR, Pertumbuhan Bobot, and Kalkulator Cepat.
- Add Hospital Pen / Sekat Seleksian feature: data model (HospitalPenData), domain use case (CalculateSeparatedFlock), Firestore service, and management screen (SekatSeleksianScreen).
- Implement Feed Intake calculation use case and integration across recording details.
- Add Partial Harvest History screen and navigation from active period card.
- Standardize UI background contrast across screens and update agent skills and rules.
- Add unit and widget tests across all new components (543 passing tests).

+++ lib/features/reporting/domain/usecases/summary_calculator.dart
@@ -141,7 +141,7 @@ class SummaryCalculator {
       ..sort((a, b) => a.day.compareTo(b.day));
 
     int totalMortality = 0;
-    int totalFeedSacks = 0;
+    double totalFeedSacks = 0.0;
     for (final r in sorted) {
       totalMortality += r.mortality;
       totalFeedSacks += r.feedSack;
